refactor: simplify component file names by removing redundant prefixes
  Rename verbose component files where folder context already provides clarity:
  - ContentDetail/: Layout, TypeBadge, MetaHeader, ActionBar
  - DiscussionEditor/: Editor, Toolbar, useEditor
  - Feed/: AggregatedItem, ItemLoading, Filters
