+++ components/ContentDetail/ActionBar.tsx
