+++ components/ContentDetail/Layout.tsx
