+++ components/ContentDetail/TypeBadge.tsx
