+++ components/ContentDetail/index.ts
@@ -1,6 +1,6 @@
-export { default as ContentDetailLayout } from "./ContentDetailLayout";
-export { default as UnifiedActionBar } from "./UnifiedActionBar";
-export { default as ContentTypeBadge } from "./ContentTypeBadge";
-export { default as ContentMetaHeader } from "./ContentMetaHeader";
+export { default as ContentDetailLayout } from "./Layout";
+export { default as UnifiedActionBar } from "./ActionBar";
+export { default as ContentTypeBadge } from "./TypeBadge";
+export { default as ContentMetaHeader } from "./MetaHeader";
 export { default as SourceInfoCard } from "./SourceInfoCard";
 export { default as InlineAuthorBio } from "./InlineAuthorBio";
