+++ components/Feed/index.ts
@@ -1,4 +1,4 @@
-export { default as FeedItemAggregated } from "./FeedItemAggregated";
-export { default as FeedItemLoading } from "./FeedItemLoading";
-export { default as FeedFilters } from "./FeedFilters";
+export { default as FeedItemAggregated } from "./AggregatedItem";
+export { default as FeedItemLoading } from "./ItemLoading";
+export { default as FeedFilters } from "./Filters";
 export { default as VoteButtons } from "./VoteButtons";
